gemini-test/requirements.txt
google-genai
PyPDF2
python-dotenv
